requirements.txt
certifi==2019.6.16
chardet==3.0.4
Click==7.0
Flask==1.1.1
h5py==2.9.0
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
numpy==1.17.0
opencv-python==4.1.0.25
Pillow==6.1.0
requests==2.22.0
torch==1.1.0
torchvision==0.3.0
urllib==1.25.3
Werkzeug==0.15.5